process/requirements.txt
connexion
sqlalchemy
connexion[swagger-ui]
Flask
mysql-connector
mysql-connector-python
pymongo
